PDEfind

Documentation for PDEfind.

PDEfind.PolynomialBasis

PDEfind.TimeDerivative

PDEfind.TimeSpaceGrid1D

PDEfind.XDerivative

PDEfind.GetVariablesMatrix

PDEfind.STRidge_cascade

PDEfind.TrainSTRidge

PDEfind.all_polynomial_functions

PDEfind.all_polynomial_vectors

PDEfind.evaluate

PDEfind.lambda_sweep

PDEfind.polynomial_vectors

PDEfind.PolynomialBasis — Type

PolynomialBasis <: AbstractBasis

Basis of polynomial functions up to degree max_degree of n_variables variables.

Initialization

PolynomialBasis(
    max_degree::Int,
    n_variables::Int
)

with ...

Fields

max_degree::Int

n_variables::Int

poly_vectors::Array

poly_functions::Array{Function}

Usage

TODO

PDEfind.TimeDerivative — Type

TimeDerivative{T}

Finite difference operator for time derivative on a grid.

Initialization

TimeDerivative(
    g::TimeSpaceGrid1D; 
    order::Int = 2,
)

with order being the order of approximation of the finite difference method.

Fields

M::T

PDEfind.TimeSpaceGrid1D — Type

TimeSpaceGrid1D{T}

Dicretization grid for time + 1D space finite difference schemes.

Initialization

TimeSpaceGrid1D(time_array, space_array)

with both assumed to have constant spacing

Fields

t::AbstractArray{T}

x::AbstractArray{T}

dt::T

dx::T

nt::Int

nx::Int

N::Int

t_grid::AbstractMatrix{T}

x_grid::AbstractMatrix{T}

PDEfind.XDerivative — Type

XDerivative{T}

Finite difference operator for (1st) spatial derivative on a grid.

Initialization

XDerivative(
    g::TimeSpaceGrid1D; 
    order::Int = 2,
    boundary_condition::String = "neumann"
)

with order being the order of approximation of the finite difference method and boundary_condition either "periodic" or "neumann" (no periodicity).

Fields

M::T

PDEfind.GetVariablesMatrix — Method

GetVariablesMatrix(
    max_derivatives::Int,
    data_array::Matrix,
    g::TimeSpaceGrid1D,
    ∂x::XDerivative

)

Create a matrix (nt*nx, max_derivatives+1) containing all j-derivatives evaluated at all time-space coordinates i. Note: This is NOT the Θ matrix.

PDEfind.STRidge_cascade — Method

STRidge_cascade(
    theta::Matrix{Float64}, 
    dt_data_array_flat::Vector{Float64}, 
    MyBasis::PolynomialBasis; 
    λ::Float64 = 0.0, 
    tol::Float64 = 0.01, 
    iters::Int64 = 1,
    verbose = false
)    
Conduct sequential thresholded ridge regression. See algorithm 1 in Rudy et al.
Returns optimal ξ.

PDEfind.TrainSTRidge — Method

TrainSTRidge(
    theta::Matrix{Float64}, 
    dt_data_array_flat::Vector{Float64}, 
    poly_vectors; 
    λ::Float64 = 1e-4, 
    tol_multiplier::Float64 = 1e2,
    iters::Int64 = 1,
    verbose::Bool = true,
    cond_number_multiplier::Float64 = 1e-2,
    max_tol_iters::Int64 = 100
)
Conduct sequential thresholded ridge regression. See algorithm 2 in Rudy et al and the .pdf file attached by the author.
Returns optimal ξ.

PDEfind.all_polynomial_functions — Method

all_polynomial_functions(max_degree::Int, n_variables::Int)

Construct all polynomials up to degree max_degree of n_variables variables. Returns array of functions of n_variables variables. Only works for 5 variables or less.

PDEfind.all_polynomial_vectors — Method

all_polynomial_vectors(max_degree::Int, n_variables::Int)

Construct all polynomials up to degree max_degree of n_variables variables. Returns array of arrays of length n_variables, i-th entry representing the power of i-th variable.

PDEfind.evaluate — Method

function evaluate(basis::PolynomialBasis, variables_array)

Evaluate basis functions at different values of variables. If basis includes M functions (of n variables) and the variables_array is an N × n array, the output is a N × M array.

PDEfind.lambda_sweep — Method

function lambda_sweep(
    lambda_range,
    Θ::Matrix{Float64}, 
    dt_data_array_flat::Vector{Float64},
    poly_vectors;
    iters::Int64 = 1,
    cond_number_multiplier::Float64 = 1e-6
)

Evaluate optimal ξ at all values of `lambda_range`. Returns array of size `(length(poly_vectors), 2)` with optimal (λ, ξ_i) values for each possible polynomial.

PDEfind.polynomial_vectors — Method

polynomial_vectors(degree::Int, n_variables::Int)

Construct all polynomials of degree exactly degree of n_variables variables. Returns array of arrays of length n_variables, i-th entry representing the power of i-th variable.